Optimize advanced reconcile workflow
The former workflow is:
1. read all data in db and keep it in memory
2. do all the matching and create all the groups of lines to reconcile
3. reconcile all the groups
The main reason to do this was probably that we wanted to be sure that
all the groups were complete before calling reconcile on them. The issue
though is that when you have hundred of thousands of lines to match, you
can quickly be in a situation where nothing reconciles, because you have
to wait the entire step 2 before actually start to reconcile, any error
during the step and all is lost.

It is actually a quick operation to check if the balance is below the
write-off, so this check is now done each time a group is modified, and
the reconcile is done directly on a positive result. Most of the times,
the groups have 2 lines, so I expect this change to be more performant
in most cases. It might be less performant if most of the groups have
more than 2 lines.

A class is used because there state to keep accross different functions
and doing it inside a model was really unreadable.