[FIX] account_reconciliation_widget: Handle blue lines
When coming from previous versions of Odoo, you may have payments
directly done against the bank account. On the reconciliation widget,
they are represented as blue lines.

One possibility is to replace in all these pending entries the bank
account by the outstanding payment/receipt account, but this means to
modify past accounting that may be locked.

So this commit is restoring the ability to reconcile against these blue
lines, although this is a deprecated thing.

Things done:

- Repair the JS widget for informing correctly about the blue lines
  to reconcile.
- When having such lines to reconcile, the temporary statement line is
  removed, and the payment one is linked.
- When reverting reconciliation of the statement lines linked to
  payments, the entry is not removed, just removed the link, and a new
  entry is created for the statement line.

TT43018