[FIX] account_reconciliation_widget: Show correct currency
When reconciling bank statement lines in a foreign currency,
it shows amounts in a company's currency while the symbol
is shown of foreign currency. Wrong currency appears on counterpart
lines and open balance line.

The wrong currency symbol only effects representation in the UI.
When reconciliation is validated, it sets company's currency on
account move lines.