[IMP] account_reconciliation_widget: Performance opening reconciliation widget
There was an unneeded mapped of the initial statements that fetches a lot of
data from statement lines that are not going to be used later, so let's
remove it and optimize a bit the initial opening time.

In a customer database, we have improved the opening time from 120 seconds
to 15.

TT48753