[FIX] account_reconcile_oca: Don't apply max_amount on intermediate counterparts
Steps to reproduce:

- Have 3 or more items to reconcile. Example: 200, -350, 150.
- One of the intermediate items should be higher than the current
  running balance.
- Select all the items, and click on Action > Reconcile.

Current behavior:

The second item appears with its amount strike-throughed (and putting
the current runnning balance as the amount to reconcile).

Expected behavior:

Only strike-through the amount if the last balance is not matching
the amount.

The solution to this is to only put a maximum amount when checking
the reconciliability of the last line of the counterparts.

TT50888