[FIX] account_reconcile_model_oca: Fix reco model matching on exact_token
When the payment label is a single word, we try to match the full value and not only the alphanumeric values.
However, the matching was made on the token that was already sanitized using the [0-9a-zA-Z\s] regex.
Instead, let's match the full value.

Related to https://github.com/odoo/odoo/commit/bccc0ae0e896d58b54debae9483b8744d2b59ddc